requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
flask>=3.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
